refactor: Move visitor to public linter package
