+++ _example/plugin/customrules/enumNamesLowerSnakeCaseRule.go
@@ -2,9 +2,9 @@ package customrules
 
 import (
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
 	"github.com/yoheimuta/protolint/internal/strs"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // EnumNamesLowerSnakeCaseRule verifies that all enum names are LowerSnakeCase.
