+++ internal/addon/rules/fieldNamesLowerSnakeCaseRule.go
@@ -2,9 +2,9 @@ package rules
 
 import (
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
 	"github.com/yoheimuta/protolint/internal/strs"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // FieldNamesLowerSnakeCaseRule verifies that all field names are underscore_separated_names.
