+++ internal/addon/rules/fieldsHaveCommentRule.go
@@ -2,8 +2,8 @@ package rules
 
 import (
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // FieldsHaveCommentRule verifies that all fields have a comment.
