+++ internal/addon/rules/importsSortedRule.go
@@ -4,9 +4,9 @@ import (
 	"sort"
 
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
 	"github.com/yoheimuta/protolint/internal/osutil"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // ImportsSortedRule enforces sorted imports.
