+++ internal/addon/rules/messagesHaveCommentRule.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // MessagesHaveCommentRule verifies that all messages have a comment.
