+++ internal/addon/rules/packageNameLowerCaseRule.go
@@ -4,9 +4,9 @@ import (
 	"strings"
 
 	"github.com/yoheimuta/go-protoparser/parser"
-	"github.com/yoheimuta/protolint/internal/addon/rules/lib/visitor"
 	"github.com/yoheimuta/protolint/internal/linter/report"
 	"github.com/yoheimuta/protolint/internal/strs"
+	"github.com/yoheimuta/protolint/linter/visitor"
 )
 
 // PackageNameLowerCaseRule verifies that the package name only contains lowercase letters, digits and/or periods.
