+++ linter/visitor/hasExtendedVisitor.go
