+++ linter/visitor/internal/disablerule/command.go
