+++ linter/visitor/internal/disablerule/interpreter.go
